separate styles from button components
