+++ src/components/styles/buttons/ButtonLink.styles.js
@@ -1,0 +1,61 @@
+import { makeStyles } from '@material-ui/core'
+import mainTheme from '../../../theme'
+
+const useStyles = makeStyles((theme) => ({
+  root: {},
+  btnWrapper: {
+    margin: '2rem 0 5rem 0',
+  },
+  button: {
+    background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)',
+    color: mainTheme.palette.defaultLight.main,
+  },
+  btn: {
+    border: 'none',
+    fontWeight: 600,
+    textDecoration: 'underline',
+    '&:hover': {
+      backgroundColor: 'transparent',
+    },
+  },
+  BtnIcon: {
+    color: ' #FE6B8B',
+  },
+  modal: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  paper: {
+    backgroundColor: theme.palette.background.paper,
+    border: '2px solid #000',
+    boxShadow: theme.shadows[5],
+    padding: theme.spacing(2, 4, 3),
+    width: '30%',
+    borderRadius: '2%',
+  },
+  form: {
+    '& .MuiTextField-root': {
+      margin: theme.spacing(1, 0),
+      [theme.breakpoints.down('xs')]: {
+        margin: theme.spacing(1),
+      },
+    },
+  },
+  textField: {
+    paddingRight: theme.spacing(1),
+  },
+  formHeader: {
+    borderBottom: '1.5px solid black',
+  },
+  submitBtn: {
+    padding: '.5rem 4rem',
+    display: 'flex',
+    marginLeft: 'auto',
+    [theme.breakpoints.down('xs')]: {
+      margin: '1rem 4.5rem',
+    },
+  },
+}))
+
+export { useStyles }
